backend/requirements.txt
fastapi
uvicorn
langchain
langchain-mistralai
langchain-community
chromadb
python-dotenv
pydantic
